Update settings to latest version of gitlab api.
According to the documentation there is many new settings available.
This commit update all settings and documentation to the latest version.
It's actually a work in progress as there is few TODOs in the code.

Basically the first verification to do is to validate that this modification is retro-compatible
with olders version of library.