Fix #658 : Pager not working properly
Since the removal of "X-Total-Pages" and "X-Total" headers, all paginations use the kaminari counter.
In this mode, kaminari counter is reset to -1 during the call of the last page. This part is correct.
Then when we call the `current()` method of the pager to get the current page. This method call the same `page` method than the `next` method. In this method the pager check if there is element before checking if we call for the current page. It throws NoSuchElementException erroneously.

To fix, I change the check sequence. First I check ifthe user call the current page. If so, we return the corresponding items. If not we check if there is such elements and call the API.