Restore API compatibility with 5.5.0 and earlier releases
https://github.com/gitlab4j/gitlab4j-api/commit/224f4c5b9f15d54ad3efa6702009e6d99c0445fd#commitcomment-145297850
notes that the change from boolean to Boolean is a breaking change in
the API.

It seems easiest and safest to restore API compatibility rather than
requiring lock-step upgrades from consumers of the API.  Jenkins (as an
example) allows users to upgrade the GitLab API plugin without requiring
that they upgrade the GitLab branch source plugin.  When that happens,
the breaking change in the API causes runtime failures for Jenkins users.

https://github.com/jenkinsci/gitlab-branch-source-plugin/issues/435
reports some of those runtime failures.

https://issues.jenkins.io/browse/JENKINS-73672 reports more runtime
failures due to the breaking change in the API.

Fixes https://github.com/gitlab4j/gitlab4j-api/issues/1155