NotesApi: fix createMergeRequestNote() API breakage introduced in 5.8.0
Even when the new `created_at` and `internal` arguments are marked as
"optional", that doesn't mean one can simply omit them from argument
list. This means `createMergeRequestNote()`'s function signature has
been changed, which is a breaking change which happens without major
version bump.

So add a wrapper function with the old signature that calls the new
signature, which restores compatibility to un-migrated code.

IMO this change is not needed for 6.x.x line, as that can be seen as
major update and API breakage is fine that way. Although, I'll admit
that with 5.8.0 released over a year ago, I'm not sure if this still
matters anymore...

Fixes: 60259a69da ("Add attributes "created_at", "internal" for NotesApi commits and DiscussionApi merge requests (#1194)")